requirements.txt
Pillow >= 8.0.0
matplotlib >= 3.3.2
selenium >= 3.141.0
opencv-python >= 4.4.0
tensorflow >= 2.3.1
keras >= 2.4.3
numpy~=1.19.5
h5py~=3.1.0

scikit-learn~=0.24.0
requests~=2.25.1
beautifulsoup4~=4.9.3
